risk_service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pandas
scikit-learn
xgboost
httpx
redis
requests